requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.1.0
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.6.1
cryptography==43.0.0
distlib==0.3.8
Django==5.0.7
django-cors-headers==4.4.0
django-sslserver==0.22
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
filelock==3.15.4
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==23.0.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jwt==1.3.1
magic-filter==1.0.12
Markdown==3.6
MarkupSafe==2.1.5
multidict==6.0.5
numpy==2.1.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
passlib==1.7.4
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-mock==1.12.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlparse==0.5.1
starlette==0.37.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
virtualenv==20.26.3
Werkzeug==3.0.3
yarl==1.9.4
